the start of a function to load in crns into a df, which will be tested for correlation with climate variable.
In order to be useful for those who are downloading crns from the ITRDB, the "include" argument is to indicate which chronologies should be included (standard, arstan, residual, earlwood, latewood, others?). These are indicated with a single letter (a, r, e, l) at the end of the file name when downloaded.