The json files contain northermost/southernmost lats, easternmost/westernmost lons, and min/max elevation. These are usually the same, because it's a single point.
Added an argument called center, which defaults to TRUE. This takes the average of those values and returns the dataframe with just lat, lon, and elevation, with the min/max columns removed.

If the min/max values aren't the same, it will be the midpoint between them, so site center.